migrate job_helpers test
